Refactored "CIS" components into "Nested"
CyOTE and possibly other modules will be using the same components, so a less specific naming is appropriate.
